V2 model version with stricter regex

+++ ai-module/app/schemas/ocr.py
@@ -12,7 +12,7 @@ except ImportError:
 class OcrExpenseItem(BaseModel):
     productName: Optional[str] = Field(None)
     price: Optional[float] = None
-    quantity: Optional[int] = 1
+    quantity: Optional[float] = 1.0
     category: Optional[str] = None
 
     model_config = {"populate_by_name": True}
